_NL2SQL_Studio/UI/requirements.txt
absl-py==2.0.0
aenum==3.1.15
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
asn1crypto==1.5.1
astroid==2.15.8
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
black==23.10.0
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
chex==0.1.83
chroma-hnswlib==0.7.3
chromadb==0.4.15
cleo==2.1.0
click==8.1.7
cloud-sql-python-connector==1.9.1
coloredlogs==15.0.1
comm==0.2.0
crashtest==0.4.1
cryptography==42.0.7
Cython==3.0.5
dataclasses-json==0.5.9
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distlib==0.3.8
docker==6.1.3
docstring_parser==0.16
dulwich==0.21.7
et-xmlfile==1.1.0
etils==1.5.1
executing==2.0.0
faiss-cpu==1.8.0
fastapi==0.104.0
fastjsonschema==2.19.1
filelock==3.12.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==23.5.26
flax==0.7.4
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.40
google-ai-generativelanguage==0.2.0
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud==0.34.0
google-cloud-aiplatform==1.50.0
google-cloud-bigquery==3.12.0
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.3.3
google-cloud-firestore==2.16.0
google-cloud-resource-manager==1.10.4
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.1.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
gspread==5.11.3
gunicorn==22.0.0
h11==0.14.0
hnswlib==0.7.0
httplib2==0.22.0
httptools==0.6.1
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
installer==0.7.0
ipykernel==6.26.0
ipython==8.16.1
isort==5.12.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jax==0.4.18
jaxlib==0.4.18
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.1
kubernetes==29.0.0
langchain==0.0.334
langchain-core==0.1.52
langchain-google-vertexai==1.0.3
langsmith==0.1.54
lazy-object-proxy==1.9.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.1
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy==1.6.1
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
networkx-stubs==0.0.1
-e git+sso://team/nl2sql/NL2SQL@f188b1a30f50d30841d17aa5f36cfc53d4abf2f0#egg=nl2sql
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.1
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.1
orjson==3.10.3
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandas-stubs==2.1.1.230928
parso==0.8.3
pathspec==0.11.2
pathvalidate==3.2.0
pexpect==4.8.0
pg8000==1.31.2
pickleshare==0.7.5
pillow==10.3.0
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
posthog==3.0.2
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.9.0
Pygments==2.16.1
pygount==1.6.1
PyJWT==2.8.0
pylint==2.17.7
pylint-protobuf==0.20.2
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.9.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.4
scramp==1.4.5
SecretStorage==3.3.3
sentence-transformers==2.7.0
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlalchemy-bigquery==1.8.0
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a37
sqlglot==19.6.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.34.0
streamlit-feedback==0.1.3
streamlit-modal==0.1.2
sympy==1.12
tenacity==8.2.3
tensorstore==0.1.46
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
torch==2.3.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.40.2
triton==2.3.0
trove-classifiers==2024.4.10
typer==0.9.0
types-openpyxl==3.1.0.27
types-protobuf==4.25.0.20240417
types-psutil==5.9.5.20240106
types-pytz==2023.3.1.1
types-requests==2.31.0.20240106
types-tqdm==4.66.0.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.18.0
virtualenv==20.26.1
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.8
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
